+++ src/Databases/Sqlite.php
@@ -4,17 +4,28 @@ declare(strict_types=1);
 
 namespace Wnx\LaravelBackupRestore\Databases;
 
+use Illuminate\Support\Facades\File;
+use Wnx\LaravelBackupRestore\Exceptions\ImportFailed;
+
 class Sqlite extends DbImporter
 {
+    /**
+     * @throws ImportFailed
+     */
     public function getImportCommand(string $dumpFile, string $connection): string
     {
-        // @todo: Improve detection of compressed files
-        if (str($dumpFile)->endsWith('gz')) {
-            // Shell command to import a gzipped SQL file to a sqlite database
-            return 'gunzip -c '.$dumpFile.' | sqlite3 '.config("database.connections.{$connection}.database");
+        if (str($dumpFile)->endsWith('sql')) {
+            return 'sqlite3 '.config("database.connections.{$connection}.database").' < '.$dumpFile;
         }
 
-        return 'sqlite3 '.config("database.connections.{$connection}.database").' < '.$dumpFile;
+        // @todo: Improve detection of compressed files
+        $decompressCommand = match (File::extension($dumpFile)) {
+            'gz' => "gunzip -c {$dumpFile}",
+            'bz2' => "bunzip2 -c {$dumpFile}",
+            default => throw ImportFailed::decompressionFailed($dumpFile, 'Unknown compression format'),
+        };
+
+        return "$decompressCommand | sqlite3 ".config("database.connections.{$connection}.database");
     }
 
     public function getCliName(): string